stage: devide to the pockets. still ne zapockecheno
