+++ auth/views.py
@@ -1,0 +1,18 @@
+from django.contrib.auth import authenticate, login
+from django.contrib.auth.forms import UserCreationForm
+from django.http import HttpResponseRedirect
+from django.views.generic.edit import CreateView
+
+from .forms import RegisterForm
+
+
+def register(request):
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            form.save()
+            new_user = authenticate(username=form.cleaned_data['username'], password=form.cleaned_data['password1'])
+            login(request, new_user)
+            return HttpResponseRedirect("/")
+    return CreateView.as_view(template_name='registration/register.html', form_class=RegisterForm,
+                              success_url='/accounts/login/')(request)
