+++ redwing/settings.py
@@ -31,7 +31,6 @@ ALLOWED_HOSTS = []
 INSTALLED_APPS = [
     'news',
     'ckeditor',
-
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
