+++ static/news/js/main.js
